Add files via upload
This is an internal file intended for testing various codes. It contains information on the households studied. Some codes use the time-step approach, while others focus on temperature evolution. Therefore, this file may not be compatible with all codes but applies to several. The data it provides on the studied households is particularly valuable.